+++ prog2-r1cs.jsonl
@@ -1,0 +1,2 @@
+{"extension_degree":1,"field_characteristic":"21888242871839275222246405745257275088548364400416034343698204186575808495617","input_variables":[0],"n_constraints":1,"n_variables":2,"output_variables":[1]}
+{"A":[["1",0]],"B":[["55",0],["11",1],["21888242871839275222246405745257275088548364400416034343698204186575808495616",2]],"C":[["0",0]]}
