+++ prog2-witness.jsonl
@@ -1,0 +1,3 @@
+{"extension_degree":1,"field_characteristic":"21888242871839275222246405745257275088548364400416034343698204186575808495617","input_variables":[0],"n_constraints":1,"n_variables":2,"output_variables":[1]}
+[0,"1"]
+[1,"66"]
